remove rotation on cards
